Disabled FreqCorrection for testing
According to the SX1276 manual the FEI Register is only valid during receive of the preamble. The PreambleDetect interrupt should be used to get good frequency offset measurements. The current implementation however just calls the Register in the Tock ISR, so there is no preamble coming in at that moment so this routine could potentially result in a wrong "correction"